backend/requirements.txt
Django==5.0.4
djangorestframework>=3.13,<4.0
psycopg2-binary==2.9.9
django-cors-headers>=4.3.1
# arize-phoenix==3.24.0
# daphne==4.1.2
# Django==5.0.4
# graphene-django==3.2.1
# python-keycloak==3.9.1
# django-cors-headers>=4.3.1
# llama-index==0.10.13
# llama-index-core==0.10.13
# llama-index-llms-azure-openai==0.1.4
# llama-index-vector-stores-qdrant==0.1.3
# llama-index-embeddings-huggingface==0.1.4
# llama-index-callbacks-arize-phoenix==0.1.4
# qdrant_client==1.7.3
# tiktoken==0.5.2
# nest_asyncio==1.5.8
# ipython
# pandas
# psycopg2-binary==2.9.9
# openai==1.2.1
# sentence-transformers==2.2.2 
# uvicorn==0.27.0